Add --in-place to rcode; harden generator tools
rcode encrypt/decrypt gain --in-place <file>: transform a file and write the
result back atomically via a sibling temp file (mkstemp) + rename, so a
mid-stream failure never corrupts the original and leaves no temp behind
(atexit cleanup). Original permission bits are preserved (fstat on the open
fd, avoiding a TOCTOU). Mutually exclusive with --in/--out.

Code-review fixes to the generator tools:
- cpscommon.h: check snprintf return in cps_build_name/cps_join so an
  over-long --out-dir/name fails loudly instead of silently truncating to a
  wrong path; widen the file-index parameter from int to size_t (%zu) to avoid
  overflow with very large --files/split counts; cps_close_out now always
  fclose()s even when fflush fails and reports the saved errno.

Tests:
- run_tests.sh: assert --in-place output is byte-identical to the --in/--out
  path (7 -> 8 in-place cases; 68 total).
- gen_tests.sh: make the N>9 space-separation test self-contained (explicit
  N=11) and also feed the pattern back through rcode; add ptn-gen --files 0;
  use sort -V for split/combine ordering (40 total).

Builds clean under -Wall -Wextra -Wpedantic -Wconversion -Wsign-conversion
-Wshadow; all tests pass.

+++ tools/ptn_gen.c
@@ -136,7 +136,7 @@ int main(int argc, char **argv) {
         if (outfile) {
             path = outfile;
         } else {
-            cps_build_name(namebuf, sizeof namebuf, ts, ".ptn", multi, (int)(fi + 1));
+            cps_build_name(namebuf, sizeof namebuf, ts, ".ptn", multi, (size_t)(fi + 1));
             cps_join(pathbuf, sizeof pathbuf, outdir, namebuf);
             path = pathbuf;
         }
